refactored marks

+++ src/main/java/pl/schoolmanagementsystem/common/mark/Mark.java
@@ -6,6 +6,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import java.math.BigDecimal;
 
 @Entity
 @Getter
@@ -19,7 +20,7 @@ public class Mark {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
-    private double mark;
+    private BigDecimal mark;
 
     private long studentId;
