+++ src/test/java/pl/schoolmanagementsystem/common/mark/MarkSamples.java
@@ -2,6 +2,8 @@ package pl.schoolmanagementsystem.common.mark;
 
 import pl.schoolmanagementsystem.common.mark.dto.MarkDto;
 
+import java.math.BigDecimal;
+
 public interface MarkSamples {
 
     String SUBJECT = "Biology";
@@ -10,7 +12,7 @@ public interface MarkSamples {
     default Mark createMark() {
         return Mark.builder()
                 .id(1)
-                .mark(4.0)
+                .mark(BigDecimal.valueOf(4.0))
                 .studentId(3)
                 .build();
     }
@@ -18,16 +20,16 @@ public interface MarkSamples {
     default Mark createMark2() {
         return Mark.builder()
                 .id(2)
-                .mark(2.0)
+                .mark(BigDecimal.valueOf(2.0))
                 .studentId(3)
                 .build();
     }
 
     default MarkDto createMarkDto1() {
-        return new MarkDto(4.0, SUBJECT);
+        return new MarkDto(BigDecimal.valueOf(4.0), SUBJECT);
     }
 
     default MarkDto createMarkDto2() {
-        return new MarkDto(2.0, SUBJECT_2);
+        return new MarkDto(BigDecimal.valueOf(2.0), SUBJECT_2);
     }
 }
